Dev/login1

+++ be/src/main/java/singleproject/bulletinboard/controller/BoardController.java
@@ -44,12 +44,6 @@ public class BoardController {
 		return "page/board_list";
 	}
 
-	@GetMapping("/board/article/write")
-	public String writeArticleForm() {
-
-		return "page/article_write";
-	}
-
 	@GetMapping("/board/article/{id}")
 	public String viewArticle(@PathVariable Long id, Model model) {
 
@@ -61,25 +55,53 @@ public class BoardController {
 		return "page/article";
 	}
 
+	@GetMapping("/board/article/write")
+	public String writeArticleForm(
+		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser) {
+
+		// 로그인 유저가 아니면 글을 작성할 수 없다.
+		if (loginUser == null) {
+			log.info("로그인을 하지 않으면 글을 작성할 수 없습니다.");
+			return "redirect:/board";
+		}
+
+		return "page/article_write";
+	}
+
 	@PostMapping("/board/article/create")
 	public String createArticle(@Valid @ModelAttribute RequestCreatedArticleInfo createdArticleInfo,
 		BindingResult bindingResult,
-		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser, RedirectAttributes redirectAttributes) {
+		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser,
+		RedirectAttributes redirectAttributes) {
 
 		if (bindingResult.hasErrors()) {
 			log.info("errors  = {}", bindingResult);
 			return "board/article_write";
 		}
 
-		Long articleId = boardService.write(createdArticleInfo.getTitle(), createdArticleInfo.getContent(),
+		// 로그인 유저가 아니면 글을 작성할 수 없다.
+		if (loginUser == null) {
+			log.info("로그인을 하지 않으면 글을 작성할 수 없습니다.");
+			return "redirect:/board";
+		}
+
+		Long articleId = boardService.write(createdArticleInfo.getTitle(),
+			createdArticleInfo.getContent(),
 			loginUser.getId());
-		redirectAttributes.addAttribute("articleId",articleId);
+		redirectAttributes.addAttribute("articleId", articleId);
 
 		return "redirect:/board/article/{articleId}";
 	}
 
 	@GetMapping("/board/{articleId}/edit")
-	public String editForm(@PathVariable Long articleId, Model model) {
+	public String editForm(@PathVariable Long articleId, Model model,
+		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser) {
+
+		if (boardService.checkWriterName(loginUser.getName(), articleId) == false) {
+			log.info("해당 글을 수정할 권한이 없습니다.");
+
+			return "redirect:/board/article/{articleId}";
+		}
 		Article article = boardService.findById(articleId).orElseThrow();
 
 		RequestCreatedArticleInfo articleForm = new RequestCreatedArticleInfo();
@@ -94,11 +116,18 @@ public class BoardController {
 
 	@PostMapping("/board/{articleId}/edit")
 	public String edit(@PathVariable Long articleId,
-		@Valid @ModelAttribute RequestCreatedArticleInfo articleForm, BindingResult bindingResult) {
+		@Valid @ModelAttribute RequestCreatedArticleInfo articleForm, BindingResult bindingResult,
+		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser) {
 
-		if(bindingResult.hasErrors()){
+		if (bindingResult.hasErrors()) {
 			log.info("errors = {}", bindingResult);
 			return "page/editForm";
+		}
+
+		if (boardService.checkWriterName(loginUser.getName(), articleId) == false) {
+			log.info("해당 글을 수정할 권한이 없습니다.");
+
+			return "redirect:/board/article/{articleId}";
 		}
 
 		boardService.updateArticle(articleId, articleForm.getTitle(), articleForm.getContent());
@@ -106,8 +135,16 @@ public class BoardController {
 		return "redirect:/board/article/{articleId}";
 	}
 
-	@GetMapping("/board/{articleId}/delete")
-	public String delete(@PathVariable Long articleId) {
+	@PostMapping("/board/{articleId}/delete")
+	public String delete(@PathVariable Long articleId,
+		@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) Member loginUser) {
+
+		if (boardService.checkWriterName(loginUser.getName(), articleId) == false) {
+			log.info("해당 글을 삭제할 권한이 없습니다.");
+
+			return "redirect:/board/article/{articleId}";
+		}
+
 		boardService.deleteById(articleId);
 		return "redirect:/board";
 	}
